core: move TravelledPath with Path and comment their (expected) use
